Polished AOP question about JoinPoint argument
